PLANET-7416 Add carousel layout to Posts List block
This is to follow designs
