+++ assets/src/editorIndex.js
@@ -1,13 +1,13 @@
 import {registerPostsListBlock} from './blocks/PostsList';
-import {registerActionsList} from './blocks/ActionsList';
+import {registerActionsListBlock} from './blocks/ActionsList';
 import {registerSubmenuBlock} from './blocks/Submenu/SubmenuBlock';
 import {registerTakeActionBoxoutBlock} from './blocks/TakeActionBoxout/TakeActionBoxoutBlock';
 import {registerHappypointBlock} from './blocks/Happypoint/HappypointBlock';
 import {setupCustomSidebar} from './block-editor/setupCustomSidebar';
-import {setupQueryBlockExtension} from './block-editor/QueryBlockExtension';
+import {setupQueryLoopBlockExtension} from './block-editor/QueryLoopBlockExtension';
 
 wp.domReady(() => {
-  // Make sure to unregister the posts-list native variation before registering planet4-blocks/posts-list-block
+  // Make sure to unregister the posts-list native variation before registering planet4-blocks/posts-list
   wp.blocks.unregisterBlockVariation('core/query', 'posts-list');
 
   // Blocks
@@ -16,7 +16,7 @@ wp.domReady(() => {
   registerHappypointBlock();
 
   // Beta blocks
-  registerActionsList();
+  registerActionsListBlock();
   registerPostsListBlock();
 });
 
@@ -24,4 +24,4 @@ setupCustomSidebar();
 
 // Setup new attributes to the core/query.
 // It should be executed after the DOM is ready
-setupQueryBlockExtension();
+setupQueryLoopBlockExtension();
